Completely get rid of all telescope associations
