+++ src/lib.rs
@@ -2,6 +2,7 @@
 #![feature(str_split_remainder)]
 pub mod action;
 pub mod app;
+pub mod app_context;
 pub mod app_input_machine;
 pub mod command;
 pub mod components;
@@ -15,9 +16,5 @@ pub mod line_entry;
 pub mod logging;
 pub mod mode;
 pub mod plugin;
-// pub mod plugin_manifest;
-pub mod simple_input_machine;
-pub mod telescope;
-pub mod telescope_query;
 pub mod themes;
 pub mod tools;
